hot fix
fix

+++ Assets/Game/Game Object/MainObjectScript.cs
@@ -18,9 +18,9 @@ public class MainObjectScript : MonoBehaviour {
   //object를 저장하기 위한 HashSet;
   private static HashSet<GameObject> game_objects_ = new HashSet<GameObject>();
   //Object의 기본 최대 체력.
-  private int max_helth_point_ = 100;
+  private float max_helth_point_ = 100.0F;
   //Object의 현재 체력.
-  public int current_helth_point_;
+  public float current_helth_point_;
   //Object의 체력이 전부 떨어졌는지 판단.
   private bool is_under_attack_ = false;
   //Object의 체력을 나타낼 체력 게이지.
@@ -39,7 +39,7 @@ public class MainObjectScript : MonoBehaviour {
   }
 
   //매개 변수로 전달된 피해량만큼 체력을 감소시킴.
-  private void DecreaseHelthPoint(int damage) {
+  private void DecreaseHelthPoint(float damage) {
     current_helth_point_ -= damage;
     //만일 체력이 모두 소진됐을 경우 Object 제거를 위한 함수 호출.
     if (current_helth_point_ <= 0) {
